requirements.txt
# Discord API
discord.py>=2.3.0
discord-py-slash-command>=4.2.1
py-cord>=2.0.0

# HTTP and API
aiohttp>=3.8.4
requests>=2.28.2

# Data processing
numpy>=1.24.3
pandas>=2.0.0

# Visualization
matplotlib>=3.7.1

# Utilities
python-dotenv>=1.0.0
typing-extensions>=4.5.0

# Audio Processing
PyNaCl>=1.4.0
wavelink>=2.0.0
yt-dlp>=2023.3.4
spotipy>=2.22.1
ffmpeg-python>=0.2.0

# Async Support
asyncio>=3.4.3